Workaround int32_t returned value from json-tools. Now frequency could be higher than 2G

+++ tools/gvsoc/common/engine/include/vp/clock/clock_engine.hpp
@@ -44,7 +44,7 @@ namespace vp {
 
     bool dequeue_from_engine();
 
-    void apply_frequency(int frequency);
+    void apply_frequency(int64_t frequency);
     
     inline clock_event *reenqueue(clock_event *event, int64_t cycles);
